+++ src/allotropy/parsers/appbio_quantstudio_designandanalysis/appbio_quantstudio_designandanalysis_contents.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import numpy as np
 import pandas as pd
 
+from allotropy.exceptions import AllotropeConversionError
 from allotropy.parsers.utils.values import (
     assert_not_empty_df,
     assert_not_none,
@@ -17,9 +18,15 @@ class DesignQuantstudioContents:
 
         self.header = self._get_header(contents)
         self.data = self._get_data(
-            self.header.size + 1,  # plus 1 for empty line between header and data
             contents,
         )
+
+    def _get_header_size(self, sheet: pd.DataFrame) -> int:
+        for idx, * (title, *_) in sheet.itertuples():
+            if title is None:
+                return int(idx)
+        error = "Unable to parse data header"
+        raise AllotropeConversionError(error)
 
     def _get_header(self, contents: dict[str, pd.DataFrame]) -> pd.Series[str]:
         sheet = assert_not_none(
@@ -27,22 +34,22 @@ class DesignQuantstudioContents:
             msg="Unable to find 'Results' sheet.",
         )
 
+        header_size = self._get_header_size(sheet)
+
         data = {}
-        for _, * (title, value, *_) in sheet.itertuples():
-            if title is None:
-                break
+        for _, * (title, value, *_) in sheet.iloc[:header_size].itertuples():
             data[str(title)] = None if value is None else str(value)
 
         header = pd.Series(data)
         header.index = header.index.str.strip()
         return header
 
-    def _get_data(
-        self, drop_n_columns: int, contents: dict[str, pd.DataFrame]
-    ) -> dict[str, pd.DataFrame]:
+    def _get_data(self, contents: dict[str, pd.DataFrame]) -> dict[str, pd.DataFrame]:
         data_structure = {}
         for name, sheet in contents.items():
-            data = sheet.iloc[drop_n_columns:].reset_index(drop=True)
+            # header_size + 1 for empty line between header and data
+            header_size = self._get_header_size(sheet) + 1
+            data = sheet.iloc[header_size:].reset_index(drop=True)
             data.columns = pd.Index(data.iloc[0])
             data_structure[name] = data.drop(0)
         return data_structure