+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed missing genotyping determination result in AppBio Quantstudio Design and Analysis
+
 ### Changed
 
 ### Deprecated