+++ CHANGELOG.md
@@ -11,6 +11,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Security
+
+## [0.1.34] - 2024-06-04
+
+### Added
+
+### Fixed
+
 - Fixed missing genotyping determination result in AppBio Quantstudio Design and Analysis
 
 ### Changed